requirements.txt
asgiref==3.7.2
black==23.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==0.7
dj-database-url==2.1.0
Django==4.2.4
django-autoslug==1.9.9
django-background-tasks-updated==1.2.7
django-compat==1.0.15
django-crispy-forms==2.0
django-environ==0.10.0
django-fast-ratelimit==0.12.0
django-filter==23.2
django-htmx==1.16.0
django-rosetta==0.9.9
django-tables2==2.6.0
flake8==6.1.0
h3==3.7.6
idna==3.4
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.25.2
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
polib==1.2.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
timezonefinder==6.2.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
whitenoise==6.5.0
